requirements.txt
rich
pandas
questionary
fpdf
openpyxl
prompt_toolkit
pillow
